Update patch set 1
Patch Set 1:

Thanks.  That example is for a committer, and Steven is already a committer in the TSC section with associated approval.  Should I list him twice with one Addition entry linking to his committer approval and another one linking to his PTL vote?   Or should I just replace the link on his current tsc entry and have it point to his PTL approval?

Patch-set: 1